Python/load_to_lake/requirements.txt
adlfs==2024.2.0
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.1
azure-datalake-store==0.0.53
azure-identity==1.15.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
frozenlist==1.4.1
fsspec==2024.2.0
idna==3.6
isodate==0.6.1
msal==1.27.0
msal-extensions==1.1.0
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==2.2.1
portalocker==2.8.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
